refact: split settings tests
