--- feature-settings/src/androidTest/java/com/jeanbarrossilva/ongoing/feature/settings/category/ActivitiesSettingsTests.kt
@@ -1,3 +1,0 @@
-package com.jeanbarrossilva.ongoing.feature.settings.category
-
-internal interface ActivitiesSettingsTests
